# openapi.yaml
openapi: 3.0.3
info:
  title: PixelProbe API
  # Version is dynamically set from version.py when served via /api/openapi.yaml
  version: 0.0.0
  description: |
    # PixelProbe REST API

    PixelProbe is a comprehensive media file corruption detection system that uses FFmpeg and other tools to scan and validate media files.

    Route and method coverage is checked in CI. Some response fields remain
    intentionally partial; consult docs/api.md and the response schemas before
    generating a strict client.

    ## Getting Started

    ### First-Time Setup

    1. Check if setup is required:
       ```bash
       curl -X GET http://your-server/api/auth/status
       ```

    2. If `first_run: true`, create admin account:
       ```bash
       curl -X POST http://your-server/api/auth/setup \
         -H "Content-Type: application/json" \
         -d '{"password": "your-secure-password"}'
       ```

    3. Login to get session cookie:
       ```bash
       curl -X POST http://your-server/api/auth/login \
         -H "Content-Type: application/json" \
         -d '{"username": "admin", "password": "your-secure-password"}' \
         -c cookies.txt
       ```

    ### Authentication Overview

    PixelProbe supports two authentication methods:

    #### Session-Based Authentication
    Used by the web interface. Login via `/auth/login` to receive a session cookie.

    #### API Token Authentication
    Recommended for programmatic access:

    1. Create an API token:
...

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
  - sessionAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Bearer token authentication using API tokens (opaque tokens generated by PixelProbe, not JWTs)
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
      description: Session-based authentication via login
...
paths:
...
  /auth/login:
    post:
      summary: User login
      description: |
        Authenticate with username and password to receive a session cookie.

        **Example:**
        ```bash
        curl -X POST http://your-server/api/auth/login \
          -H "Content-Type: application/json" \
          -c cookies.txt \
          -d '{
            "username": "admin",
            "password": "your-password",
            "remember": true
          }'
        ```
      tags: [Authentication]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
                remember:
                  type: boolean
                  default: false
      responses:
        '200':
          description: Login successful
        '401':
          description: Invalid credentials
...
  /auth/setup:
    post:
      summary: First-run setup
      description: |
        Initialize PixelProbe with the admin password on first run.

        **Example:**
        ```bash
        curl -X POST http://your-server/api/auth/setup \
          -H "Content-Type: application/json" \
          -d '{"password": "your-secure-password"}'
        ```
      tags: [Authentication]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password:
                  type: string
      responses:
        '200':
          description: Setup completed successfully
        '400':
          description: Invalid request or setup already completed
  /users:
    get:
      summary: List all users
      description: |
        Retrieve a list of all user accounts (admin only).

        **Example:**
        ```bash
        curl -X GET http://your-server/api/users \
          -H "Authorization: Bearer YOUR_TOKEN"
        ```
      tags: [Authentication]
      responses:
        '200':
          description: List of users retrieved successfully

    post:
      summary: Create new user
      description: |
        Create a new user account with specified credentials and permissions (admin only).

        **Example:**
        ```bash
        curl -X POST http://your-server/api/users \
          -H "Authorization: Bearer YOUR_TOKEN" \
          -H "Content-Type: application/json" \
          -d '{
            "username": "newuser",
            "email": "newuser@example.com",
            "password": "secure-password",
            "is_admin": false
          }'
        ```
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, email, password]
              properties:
                username:
                  type: string
                email:
                  type: string
                password:
                  type: string
                is_admin:
                  type: boolean
                  default: false
      responses:
        '200':
          description: User created successfully
        '400':
          description: Invalid request or username already exists
  /users/{user_id}:
    delete:
      summary: Delete user
      description: |
        Delete a user account by ID (admin only).

        **Example:**
        ```bash
        curl -X DELETE http://your-server/api/users/2 \
          -H "Authorization: Bearer YOUR_TOKEN"
        ```
      tags: [Authentication]
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User deleted successfully
        '400':
          description: Cannot delete user
        '404':
          description: User not found
  /users/{user_id}/password:
    put:
      summary: Change user password
      description: |
        Change the password for a specific user account.

        **Example:**
        ```bash
        curl -X PUT http://your-server/api/users/1/password \
          -H "Authorization: Bearer YOUR_TOKEN" \
          -H "Content-Type: application/json" \
          -d '{
            "current_password": "old-password",
            "new_password": "new-secure-password"
          }'
        ```
      tags: [Authentication]
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password:
                  type: string
                new_password:
                  type: string
      responses:
        '200':
          description: Password changed successfully
        '401':
          description: Invalid current password
  /tokens:
    get:
      summary: List API tokens
      description: |
        Retrieve all API tokens for the current user.

        **Example:**
        ```bash
        curl -X GET http://your-server/api/tokens \
          -H "Authorization: Bearer YOUR_TOKEN"
        ```
      tags: [Authentication]
      responses:
        '200':
          description: List of API tokens retrieved successfully

    post:
      summary: Create API token
      description: |
        Generate a new API token for programmatic access.

        **Example:**
        ```bash
        curl -X POST http://your-server/api/tokens \
          -H "Authorization: Bearer YOUR_TOKEN" \
          -H "Content-Type: application/json" \
          -d '{
            "description": "Monitoring script",
            "expires_in_days": 365
          }'
        ```
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [description]
              properties:
                description:
                  type: string
                expires_in_days:
                  type: integer
      responses:
        '200':
          description: Token created successfully
  /tokens/{token_id}:
    delete:
      summary: Revoke API token
      description: |
        Delete and revoke an API token by ID.

        **Example:**
        ```bash
        curl -X DELETE http://your-server/api/tokens/2 \
          -H "Authorization: Bearer YOUR_TOKEN"
        ```
      tags: [Authentication]
      parameters:
        - name: token_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Token revoked successfully
        '404':
          description: Token not found

  # Scan Endpoints
...